feat: add ability to sort popular agents by popularity and recently added refactor: rename CreateWizards to ViewAgents
