+++ src/components/ViewAgents/MyWizards/DeleteWizardModal.tsx
